requirements.txt
asgiref==3.8.1
Django==5.1.2
django-admin-kubi==1.1.0
djangorestframework==3.15.2
pillow==11.0.0
six==1.16.0
sqlparse==0.5.1
whitenoise==6.8.2
